remove unneeded abstraction
