+++ vnr/vnrhook/CMakeLists.txt
@@ -60,8 +60,6 @@ set(vnrhook_src
   ${PROJECT_SOURCE_DIR}/ntinspect/ntinspect.cc
   ${PROJECT_SOURCE_DIR}/ntinspect/ntinspect.h
   ${PROJECT_SOURCE_DIR}/winkey/winkey.h
-  ${PROJECT_SOURCE_DIR}/winversion/winversion.cc
-  ${PROJECT_SOURCE_DIR}/winversion/winversion.h
   ${PROJECT_SOURCE_DIR}/mono/monoobject.h
   ${PROJECT_SOURCE_DIR}/mono/monotype.h
 )
